@@ -60,16 +60,16 @@
 		"@kitschpatrol/typescript-config": "workspace:*",
 		"@pinojs/json-colorizer": "^4.0.0",
 		"@types/fs-extra": "^11.0.4",
-		"@types/node": "^20.17.24",
-		"@types/react": "^19.0.10",
+		"@types/node": "^20.17.28",
+		"@types/react": "^19.0.12",
 		"@types/yargs": "^17.0.33",
-		"bumpp": "^10.0.3",
+		"bumpp": "^10.1.0",
 		"chalk": "^5.4.1",
 		"cosmiconfig": "^9.0.0",
 		"cosmiconfig-typescript-loader": "^6.1.0",
 		"decircular": "^1.0.0",
 		"deepmerge": "^4.3.1",
-		"esbuild": "^0.25.0",
+		"esbuild": "^0.25.1",
 		"execa": "^9.5.2",
 		"find-workspaces": "^0.3.1",
 		"fs-extra": "^11.3.0",
@@ -77,11 +77,11 @@
 		"package-up": "^5.0.0",
 		"prettier": "^3.5.3",
 		"tsx": "^4.19.3",
-		"typescript": "~5.7.3",
-		"vitest": "^3.0.8",
+		"typescript": "~5.8.2",
+		"vitest": "^3.0.9",
 		"yargs": "^17.7.2"
 	},
-	"packageManager": "pnpm@10.6.1",
+	"packageManager": "pnpm@10.7.0",
 	"engines": {
 		"node": ">=20.9.0"
 	},
