@@ -49,35 +49,35 @@
 	},
 	"dependencies": {
 		"@eslint-community/eslint-plugin-eslint-comments": "^4.4.1",
-		"@eslint-react/eslint-plugin": "^1.31.0",
-		"@html-eslint/eslint-plugin": "^0.35.2",
-		"@html-eslint/parser": "^0.35.2",
+		"@eslint-react/eslint-plugin": "^1.38.2",
+		"@html-eslint/eslint-plugin": "^0.37.0",
+		"@html-eslint/parser": "^0.37.0",
 		"@pinojs/json-colorizer": "^4.0.0",
-		"@typescript-eslint/eslint-plugin": "^8.26.0",
-		"@typescript-eslint/parser": "^8.26.0",
-		"@vitest/eslint-plugin": "^1.1.36",
-		"astro-eslint-parser": "^1.2.1",
-		"eslint": "^9.22.0",
+		"@typescript-eslint/eslint-plugin": "^8.28.0",
+		"@typescript-eslint/parser": "^8.28.0",
+		"@vitest/eslint-plugin": "^1.1.38",
+		"astro-eslint-parser": "^1.2.2",
+		"eslint": "^9.23.0",
 		"eslint-config-flat-gitignore": "^2.1.0",
 		"eslint-flat-config-utils": "^2.0.1",
-		"eslint-import-resolver-typescript": "^3.8.3",
-		"eslint-mdx": "^3.2.0",
+		"eslint-import-resolver-typescript": "^3.10.0",
+		"eslint-mdx": "^3.3.1",
 		"eslint-plugin-astro": "^1.3.1",
 		"eslint-plugin-depend": "^0.12.0",
 		"eslint-plugin-html": "^8.1.2",
-		"eslint-plugin-import-x": "^4.6.1",
-		"eslint-plugin-jsdoc": "^50.6.3",
-		"eslint-plugin-jsonc": "^2.19.1",
+		"eslint-plugin-import-x": "^4.9.3",
+		"eslint-plugin-jsdoc": "^50.6.9",
+		"eslint-plugin-jsonc": "^2.20.0",
 		"eslint-plugin-jsx-a11y": "^6.10.2",
-		"eslint-plugin-mdx": "^3.2.0",
-		"eslint-plugin-n": "^17.16.2",
+		"eslint-plugin-mdx": "^3.3.1",
+		"eslint-plugin-n": "^17.17.0",
 		"eslint-plugin-no-only-tests": "^3.3.0",
-		"eslint-plugin-package-json": "^0.26.3",
+		"eslint-plugin-package-json": "^0.29.0",
 		"eslint-plugin-perfectionist": "^4.10.1",
 		"eslint-plugin-regexp": "^2.7.0",
-		"eslint-plugin-svelte": "^3.0.3",
+		"eslint-plugin-svelte": "^3.3.3",
 		"eslint-plugin-toml": "^0.12.0",
-		"eslint-plugin-unicorn": "^57.0.0",
+		"eslint-plugin-unicorn": "^58.0.0",
 		"eslint-plugin-yml": "^1.17.0",
 		"execa": "^9.5.2",
 		"find-workspaces": "^0.3.1",
@@ -87,7 +87,7 @@
 		"local-pkg": "^1.1.1",
 		"prettier": "^3.5.3",
 		"sort-package-json": "^3.0.0",
-		"svelte-eslint-parser": "^1.0.1",
+		"svelte-eslint-parser": "^1.1.1",
 		"toml-eslint-parser": "^0.10.0",
 		"yaml-eslint-parser": "^1.3.0"
 	},
@@ -97,9 +97,9 @@
 		"dot-prop": "^9.0.0",
 		"eslint-config-prettier": "^10.1.1",
 		"eslint-config-xo-typescript": "^7.0.0",
-		"eslint-typegen": "^2.0.0",
+		"eslint-typegen": "^2.1.0",
 		"globby": "^14.1.0",
-		"svelte": "^5.22.6",
+		"svelte": "^5.25.3",
 		"tsup": "^8.4.0"
 	},
 	"engines": {
