@@ -3,7 +3,8 @@ import type { Rules } from '../types'
 // Includes all of recommended-typescript, which includes all recommended rules that aren't better handled by type-unaware typescript
 export const reactRecommendedTypeCheckedRules: Rules = {
 	// Begin expansion '@eslint-react/eslint-plugin' 'recommended-type-checked'
-	'react/ensure-forward-ref-using-ref': 'warn',
+	'react/jsx-no-duplicate-props': 'off',
+	'react/jsx-uses-vars': 'off',
 	'react/no-access-state-in-setstate': 'error',
 	'react/no-array-index-key': 'warn',
 	'react/no-children-count': 'warn',
@@ -20,12 +21,11 @@ export const reactRecommendedTypeCheckedRules: Rules = {
 	'react/no-create-ref': 'error',
 	'react/no-default-props': 'error',
 	'react/no-direct-mutation-state': 'error',
-	'react/no-duplicate-jsx-props': 'off',
 	'react/no-duplicate-key': 'warn',
 	'react/no-forward-ref': 'warn',
 	'react/no-implicit-key': 'warn',
 	'react/no-missing-key': 'error',
-	'react/no-nested-components': 'error',
+	'react/no-nested-component-definitions': 'error',
 	'react/no-prop-types': 'error',
 	'react/no-redundant-should-component-update': 'error',
 	'react/no-set-state-in-component-did-mount': 'warn',
@@ -40,28 +40,30 @@ export const reactRecommendedTypeCheckedRules: Rules = {
 	'react/no-unused-class-component-members': 'warn',
 	'react/no-unused-state': 'warn',
 	'react/no-use-context': 'warn',
-	'react/use-jsx-vars': 'off',
+	'react/no-useless-forward-ref': 'warn',
+	'react/no-useless-fragment': 'warn',
 	'react-dom/no-dangerously-set-innerhtml': 'warn',
 	'react-dom/no-dangerously-set-innerhtml-with-children': 'error',
 	'react-dom/no-find-dom-node': 'error',
+	'react-dom/no-flush-sync': 'error',
 	'react-dom/no-missing-button-type': 'warn',
 	'react-dom/no-missing-iframe-sandbox': 'warn',
 	'react-dom/no-namespace': 'error',
+	'react-dom/no-render': 'error',
 	'react-dom/no-render-return-value': 'error',
 	'react-dom/no-script-url': 'warn',
+	'react-dom/no-unknown-property': 'off',
 	'react-dom/no-unsafe-iframe-sandbox': 'warn',
 	'react-dom/no-unsafe-target-blank': 'warn',
-	'react-dom/no-void-elements-with-children': 'warn',
+	'react-dom/no-use-form-state': 'error',
+	'react-dom/no-void-elements-with-children': 'error',
 	'react-web-api/no-leaked-event-listener': 'warn',
 	'react-web-api/no-leaked-interval': 'warn',
 	'react-web-api/no-leaked-resize-observer': 'warn',
 	'react-web-api/no-leaked-timeout': 'warn',
 	'react-hooks-extra/no-direct-set-state-in-use-effect': 'warn',
-	'react-hooks-extra/no-useless-custom-hooks': 'warn',
+	'react-hooks-extra/no-unnecessary-use-prefix': 'warn',
 	'react-hooks-extra/prefer-use-state-lazy-initialization': 'warn',
-	'react-naming-convention/context-name': 'warn',
-	'react-naming-convention/use-state': 'warn',
-	'react-dom/no-unknown-property': 'off',
 	'react/no-leaked-conditional-rendering': 'warn',
 	// End expansion
 }
