@@ -1,9 +1,10 @@
 #!/usr/bin/env node
 
+import type { foregroundColorNames } from 'chalk'
+import type { CosmiconfigResult } from 'cosmiconfig'
 import type internal from 'node:stream'
-// eslint-disable-next-line unicorn/import-style
-import chalk, { type foregroundColorNames } from 'chalk'
-import { cosmiconfig, type CosmiconfigResult } from 'cosmiconfig'
+import chalk from 'chalk'
+import { cosmiconfig } from 'cosmiconfig'
 import { TypeScriptLoader as typeScriptLoader } from 'cosmiconfig-typescript-loader'
 import { execa } from 'execa'
 import fse from 'fs-extra'
@@ -14,10 +15,11 @@ import { fileURLToPath } from 'node:url'
 import { packageUp } from 'package-up'
 import yargs from 'yargs'
 import { hideBin } from 'yargs/helpers'
+import type { CwdOverrideOptions } from './path-utilities.js'
 import { version } from '../package.json'
 import { isErrorExecaError } from './execa-utilities.js'
 import { merge, stringify } from './json-utilities.js'
-import { type CwdOverrideOptions, getCwdOverride } from './path-utilities.js'
+import { getCwdOverride } from './path-utilities.js'
 import { formatFileInPlace } from './prettier-utilities.js'
 import { createStreamTransform, streamToString } from './stream-utilities.js'
 import { pluralize } from './string-utilities.js'
